feat (inventory): Se corrigió un tema central: no eliminas un producto del almacén o inventario porque agotó el stock... desaparece sí del listado, pero continúa en registros históricos, como en Inventory_Movements. AGregué state o estado en Inventory (1: En stock; 0: Agotado). Pruebas en Postman exitosas, aunque trémulo, temeroso, acepto el destino.

+++ src/main/java/com/intellimarket/api/inventory/dto/ProductsResponse.java
@@ -1,6 +1,7 @@
 package com.intellimarket.api.inventory.dto;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public record ProductsResponse (
         String name,
@@ -8,5 +9,5 @@ public record ProductsResponse (
         String description,
         Integer stock,
         BigDecimal price,
-        Long id
+        BigInteger id
 ) {}
